requirements.txt
APScheduler==3.7.0
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-email-verification==0.1.0
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
Markdown==3.3.4
oauthlib==3.1.1
Pillow==8.1.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.4
